create admin check middleware and also check only admin can create the product,delete or update.

+++ app/Http/Controllers/Admin/ProductController.php
@@ -1,47 +1,31 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Admin;
 
-use App\ApiResponseTrait;
+use App\Http\Controllers\Controller;
 use App\Http\Requests\ProductRequest;
 use App\Models\Product;
 use Exception;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
-use Psy\CodeCleaner\ReturnTypePass;
 
 class ProductController extends Controller
 {
-    use ApiResponseTrait;
+
+    /** this is only access admin */
     /**
-     * Display a listing of the resource.
+     * Store a newly created resource in storage.
      */
-    public function index()
+    public function store(ProductRequest $request)
     {
         try {
-            $products = Product::all();
-            return $this->successResponse($products, 'All Products', 200);
+            $product = Product::create($request->all());
+            return $this->successResponse($product, 'product create successfully', 201);
         } catch (Exception $e) {
-            return $this->errorResponse('server Error', 500);
+            return $this->errorResponse('Server Error', 500, $e->getMessage());
         }
     }
-
-    /**
-     * Display the specified resource.
-     */
-    public function show($id)
-    {
-        try {
-            $product = Product::findOrFail($id);
-            return $this->successResponse($product, 'Prdouct Found', 200);
-        } catch (ModelNotFoundException $e) {
-            return $this->errorResponse('Product not Found', 404);
-        } catch (Exception $e) {
-            return $this->errorResponse('Server Error', 500);
-        }
-    }
-
-    /* Update the specified resource in storage.
+     /* Update the specified resource in storage.
      */
     public function update(ProductRequest $request, $id)
     {
@@ -60,10 +44,12 @@ class ProductController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy($id)
+    public function destroy(Product $product)
     {
         try {
-            $product = Product::findOrFail($id);
+            if(!$product){
+                throw new ModelNotFoundException('Invalid Product');
+            }
             $product->delete();
             return $this->successResponse([], 'product Delete successfully', 200);
         } catch (ModelNotFoundException $e) {
